Print a basic qstat summary.
This commit adds a Job object, adds a method to create a set of jobs
objects from the output of qstat, and a handful of methods to nicely
format and display summaries of those jobs.  In its current state, the
qtop command displays all the same information as the cluster_status.sh
script, but is faster and reacts automatically to new lab members.